+++ truck-shapeops/src/transversal/mod.rs
@@ -5,4 +5,4 @@ pub(crate) mod intersection_curve;
 mod loops_store;
 mod polyline_construction;
 pub(crate) use integrate::smooth_leader;
-pub use integrate::{and, or, ShapeOpsCurve, ShapeOpsSurface};
+pub use integrate::{and, or, solid_components, subtract, ShapeOpsCurve, ShapeOpsSurface};